Fix schema integration test
I was setting the path after the command constructor
but the constructor checks if the path to the command exists and sets a
flag to fail on Start when it cannot be found. So in this case if docker
isn't in your path, it will fail even though I set the full path to the
docker mixin on the next line.

The proper way to use exec.Command is to set the full path to the
command to call in the constructor and it cannot be changed later.